requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
Django==4.2
django-embed-video==1.4.8
exceptiongroup==1.1.1
Faker==18.4.0
h11==0.14.0
idna==3.4
numpy==1.24.2
outcome==1.2.0
pandas==2.0.0
psycopg2==2.9.6
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tmdbv3api==1.7.7
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.15
wsproto==1.2.0
